requirements.txt
black==24.4.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docstring-parser==0.15
google-api-core==2.17.1
google-auth==2.29.0
google-cloud-aiplatform==1.49.0
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.1
google-cloud-storage==2.16.0
google-crc32c==1.6.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
h3==3.7.7
idna==3.7
joblib==1.4.2
kfp==2.7.0
kfp-server-api==2.0.7
kubernetes==29.0.0
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.1
shapely==2.0.4
six==1.16.0
tabulate==0.9.0
threadpoolctl==3.4.0
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.8.0
python-dotenv==1.0.0
